Additional tests capturing some gets. Also fixed a bug in getDoubleFiled that was incorrectly calling getLongfield resulting in truncation.

+++ CoreUtilities/src/au/gov/asd/tac/constellation/utilities/json/JsonUtilities.java
@@ -149,7 +149,7 @@ public class JsonUtilities {
     }
 
     public static double getDoubleField(JsonNode node, String... keys) {
-        return getLongField(0, node, keys);
+        return getDoubleField(0, node, keys);
     }
 
     public static double getDoubleField(double defaultVal, JsonNode node, String... keys) {
